tests/wpt/tests/speculation-rules/speculation-tags/prerender-target-hint.https.html

<!DOCTYPE html>
<title>Sec-Speculation-Tags request headers</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="/html/browsers/browsing-the-web/remote-context-helper/resources/remote-context-helper.js"></script>
<script src="/speculation-rules/resources/utils.js"></script>
<script src="resources/speculation-tags-utils.js"></script>
<script>
"use strict";

setup(() => assertSpeculationRulesIsSupported());

// Test that tags are sent for prerendering with target_hint.
promise_test(async t => {
    // Inputs and expectations:
    const rulesetTag = "def";
    const ruleTags = ["abc", "ghi"];
    const expectedTags = '"abc", "def", "ghi"';

    const rcHelper = new RemoteContextHelper();
    const referrerRC = await rcHelper.addWindow(undefined, { features: 'noopener' });

    const extraConfig = {};
    const preloadedRC = await referrerRC.helper.createContext({
        executorCreator(url) {
          return referrerRC.executeScript((url, rulesetTag, ruleTags) => {
              const script = document.createElement("script");
              script.type = "speculationrules";
              script.textContent = JSON.stringify({
                  tag:rulesetTag,
                  prerender: [
                    {
                      tag: ruleTags[0],
                      target_hint: "_blank",
                      source: "list",
                      urls: [url]
                    },
                    {
                      tag: ruleTags[1],
                      target_hint: "_blank",
                      source: "list",
                      urls: [url]
                    }
                  ]
              });
              document.head.append(script);
          }, [url, rulesetTag, ruleTags]);
        }, extraConfig
    });

    // Open the prerendered page in a new tab.
    await referrerRC.navigate((url) => {
        window.open(url, "_blank", "noopener");
    }, [preloadedRC.url]);

    const headers = await preloadedRC.getRequestHeaders();
    // Make sure the page is prerendered.
    assert_equals(headers.get("sec-purpose"), "prefetch;prerender");
    assert_equals(headers.get("sec-speculation-tags"), expectedTags);
}, "Sec-Speculation-Tags: tags with prerender target hint");

</script>
